Work in progress. Save all changes.
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@4041 b967ad99-d558-0410-b138-e0f6c56caec7

+++ ImgAlgos/trunk/include/CSPadArrPeakFinder.h
@@ -66,8 +66,8 @@ struct MedianResult {
 };
 
 struct Peak{
-  double quad;
-  double sect; 
+  int    quad;
+  int    sect; 
   double col;
   double row; 
   double sigma_col;
